requirements.txt
alembic==0.9.3
certifi==2018.8.24
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Cors==3.0.6
Flask-Migrate==2.0.4
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
idna==2.7
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
python-dateutil==2.6.1
python-dotenv==0.8.2
python-editor==1.0.3
requests==2.19.1
six==1.10.0
slackclient==1.1.3
SQLAlchemy==1.1.11
urllib3==1.23
websocket-client==0.53.0
Werkzeug==0.12.2
XlsxWriter==0.9.8
